refactor: clubbed similar services in one directory
