+++ src/context/UserContext/UserContext.jsx
@@ -77,7 +77,6 @@ export const UserProvider = ({ children }) => {
   };
 
   const shareVideoHandler = async (videoID) => {
-    console.log(videoID);
     await navigator.clipboard.writeText(
       `${window.location.origin}/watch/${videoID}`
     );
