+++ src/services/index.js
@@ -1,16 +1,10 @@
-export { addNewUser } from "./addNewUser";
-export { getAllVideos } from "./getAllVideos";
-export { getCategories } from "./getCategories";
-export { getAllLanguages } from "./getAllLanguages";
-export { getUser } from "./getUser";
-export { addLikedVideo } from "./addLikedVideo";
-export { removeLikedVideo } from "./removeLikedVideo";
-export { addSavedVideo } from "./addSavedVideo";
-export { removeSavedVideo } from "./removeSavedVideo";
+export { addNewUser, getUser } from "./auth";
+export { addLikedVideo, removeLikedVideo } from "./like";
+export { addSavedVideo, removeSavedVideo } from "./saved";
+export { getAllVideos, getCategories, getAllLanguages } from "./data";
 export {
   createPlaylist,
   removePlaylist,
   addVideoToPlaylist,
   removeVideoFromPlaylist,
 } from "./playlists";
-// TODO: refactor all services and club them into one file which are related as done with playlists.
